Added delete confirmations. Tinkered with edit profile, but for now it's turned off.

+++ src/app/event-pane/event-details/event-details.component.ts
@@ -52,7 +52,9 @@ export class EventDetailsComponent implements OnInit, OnDestroy {
     }
 
     onDelete() {
-        this.store.dispatch(new DeleteEvent(this.event._id));
+        if (confirm('Are you sure you want to delete your event?')) {
+            this.store.dispatch(new DeleteEvent(this.event._id));
+        }
     }
 
     onChangeParticipation() {
