+++ src/app/feed/post/post.component.ts
@@ -102,7 +102,9 @@ export class PostComponent implements OnInit, OnDestroy {
     }
 
     onDelete() {
-        this.postService.deletePost(this.post.id);
+        if (confirm('Are you sure you want to delete your post?')) {
+            this.postService.deletePost(this.post.id);
+        }
     }
 
     onNewComment() {
